Extract IPC helpers
